Replace python template with a yaml template deployed in cluster (#35)
* Replace python template with a yaml template deployed in cluster

This will allow system engineers to more easily make configurations
to the suite runner Job without adding code to the repository.
This template should become a CRD in the future when we are ready
for ETOS to become an operator.

The filebeat template is hard to use since we tend to deploy ETOS
using the etos repository. We also do not have filebeat configurations
for any other service in ETOS.
This should be up to the deployer of ETOS to set up for all services.